+++ python/cucim/src/cucim/skimage/filters/_separable_filtering.py
